requirements.txt
ipykernel==6.29.4
langchain==0.1.16
langchain-community==0.0.34
langchain-openai
openai
pypdf==4.2.0
streamlit==1.33.0
python-dotenv==1.0.1
faiss-cpu==1.8.0
PyMuPDF==1.25.3
pdf2image